server-service 컨트롤러, 서비스 리펙터링
