+++ .idea/compiler.xml
@@ -7,8 +7,31 @@
         <processorPath useClasspath="false">
           <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.projectlombok/lombok/1.18.34/ec547ef414ab1d2c040118fb9c1c265ada63af14/lombok-1.18.34.jar" />
         </processorPath>
-        <module name="Homeless_User.main" />
+        <module name="com.example.friends-server.main" />
       </profile>
     </annotationProcessing>
+    <bytecodeTargetLevel>
+      <module name="chat-service" target="17" />
+      <module name="chat-service.main" target="17" />
+      <module name="chat-service.test" target="17" />
+      <module name="config-serivce" target="17" />
+      <module name="config-serivce.main" target="17" />
+      <module name="config-serivce.test" target="17" />
+      <module name="discovery-service" target="17" />
+      <module name="discovery-service.main" target="17" />
+      <module name="discovery-service.test" target="17" />
+      <module name="friends-server" target="17" />
+      <module name="friends-server.main" target="17" />
+      <module name="friends-server.test" target="17" />
+      <module name="gateway-service" target="17" />
+      <module name="gateway-service.main" target="17" />
+      <module name="gateway-service.test" target="17" />
+      <module name="homeless-code" target="17" />
+      <module name="homeless-code.main" target="17" />
+      <module name="homeless-code.test" target="17" />
+      <module name="Homeless_User" target="17" />
+      <module name="Homeless_User.main" target="17" />
+      <module name="Homeless_User.test" target="17" />
+    </bytecodeTargetLevel>
   </component>
 </project>
