+++ server-service/src/main/java/com/playdata/homelesscode/controller/ServerController.java
@@ -101,7 +101,6 @@ public class ServerController {
     }
 
 
-
     ////////////////////////////////////////////// 서버관리 /////////////////////////////////////////////////////////////////
 
     // 서버 추가요청
