+++ user-service/src/main/java/com/spring/homeless_user/user/repository/UserRepository.java
@@ -19,4 +19,5 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     Boolean existsByEmail(String email);
 
+    List<User> findByEmailIn(List<String> userEmails);
 }
